Refactor: Move pipeline to server/lib/alert_investigation (independent of Attack Discovery)
Major architectural refactoring for minimal cross-team dependencies:

**Directory Move:**
- FROM: server/lib/attack_discovery/pipeline/
- TO: server/lib/alert_investigation/
- Flattened: alert_investigation/pipeline/* → alert_investigation/*

**Why:**
- Pipeline is conceptually independent (only 20% depends on AD)
- 4/5 stages are generic (dedup, extract, match, tag)
- Workflow steps are reusable across ANY workflow (not AD-specific)
- Enables shipping without Attack Discovery team approval

**Files moved (43 files):**
- Deduplication, entity extraction, case matching, incremental AD
- Workflow steps (fetch, dedup, extract, tag)
- Metrics, audit, validation, observables cache
- Orchestrator, types, enrichment strategies
- All tests

**Routes moved:**
- FROM: routes/attack_discovery/pipeline/
- TO: routes/alert_investigation/
- API paths unchanged (backward compatible)

**Import updates:**
- Updated 43 TypeScript files
- Fixed relative import paths (../../../ → ../../)
- plugin.ts: registerPipelineWorkflowSteps import updated
- All route imports updated

**Validation:**
- TypeScript: ✅ PASSED (0 errors)
- Structure: lib/alert_investigation/ (independent)
- Attack Discovery integration: case_integration/trigger_case_ad.ts (optional)

**Cross-team dependencies: ZERO**
- No Team:ResponseOps approval needed (independent of Agent Builder)
- No Team:DetectionEngine approval (self-contained)
- No Team:Cases approval (uses existing APIs)
- Approvals needed: elastic_assistant maintainers ONLY

**Enables shipping with 1 team approval instead of 6 teams.**
