+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/case_integration/index.ts
