+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/case_integration/trigger_case_ad.test.ts
