+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/enrichment_strategies/mitre_attack_enrichment.test.ts
