+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/entity_extraction/extract_entities.test.ts
