+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/incremental/incremental_processor.ts
