+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/task_manager/pipeline_task.ts
