+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/types.ts
