+++ x-pack/solutions/security/plugins/elastic_assistant/server/lib/alert_investigation/workflow_steps/alert_pipeline_steps.ts
