+++ x-pack/solutions/security/plugins/elastic_assistant/server/routes/register_routes.ts
@@ -59,8 +59,8 @@ import { suggestUsersRoute } from './users/suggest';
 import { updateAnonymizationFieldsRoute } from './test_internal/update_anonymization_fields_route';
 import { getMissingIndexPrivilegesInternalRoute } from './attack_discovery/privileges/get_missing_privileges';
 import { createAttackDiscoveryAlertsRoute } from './test_internal/create_attack_discovery_alerts_route';
-import { registerPipelineRoutes } from './attack_discovery/pipeline/post_pipeline_run';
-import { registerPipelineObservabilityRoutes } from './attack_discovery/pipeline/pipeline_observability';
+import { registerPipelineRoutes } from './alert_investigation/post_pipeline_run';
+import { registerPipelineObservabilityRoutes } from './alert_investigation/pipeline_observability';
 
 export const registerRoutes = (
   router: ElasticAssistantPluginRouter,
